Mergiato Convertitore in Traduttore
Spostati metodi e attributi di Convertitore.java in Traduttore.java e lo stesso per i test
