+++ gateway/src/main/java/redroundrobin/gateway/Traduttore.java
@@ -6,25 +6,46 @@ import kafka.utils.json.JsonObject;
 import java.util.ArrayList;
 
 public class Traduttore {
+    public static final int ERR = -1;
+    public static final int REQ = 0;
+    public static final int RES = 1;
 
-    //Metodo che trasforma la stringa ricevuta in input in una stringa Json (da inviare a Kafka)
-    boolean inJSON(String messaggio) {
-        //Struttura JSON1: idDevice, idSensore, dato, idSensore, dato, ... , idDevice, ...
+    private ArrayList<Dispositivo> lista;
 
-        return false;
+    public Traduttore() {
+        lista = new ArrayList<>();
     }
 
-    /*
-    * Metodo che dato un JsonObject lo trasforma in una stringa
-    */
-    boolean inStringa(JsonObject json){
-        return false;
+    public void aggiungiSensore( byte[] b ) {
+        if( b[1] == ERR || b[1] == REQ || b[1] == RES ) {
+            int id = Byte.toUnsignedInt(b[0]);
+
+            int sensore = Byte.toUnsignedInt(b[2]);
+            int dato = Byte.toUnsignedInt(b[3]);
+
+            if( lista.stream().anyMatch(d -> d.getId() == id) ) {
+                Dispositivo d = lista.stream().filter(i -> i.getId() == id).findFirst().get();
+
+                if( d.getSensori().stream().anyMatch(s -> s.getId() == sensore) ) {
+                    d.getSensori().stream().filter(s -> s.getId() == sensore).findFirst().get().setDato(dato);
+                } else {
+                    d.getSensori().add(new Sensore(sensore, dato));
+                }
+            } else {
+                Dispositivo d = new Dispositivo(id);
+                d.getSensori().add(new Sensore(sensore, dato));
+
+                lista.add(d);
+            }
+        }
     }
 
-    public String inJSON(ArrayList<Dispositivo> lista) {
+    public ArrayList<Dispositivo> getDispositivi() {
+        return this.lista;
+    }
+
+    public String getJSON() {
         Gson gson = new Gson();
-        return gson.toJson(lista);
+        return gson.toJson(this.lista);
     }
-
-
 }
